Create Wavelet algorithim for simple compression
In this example, we first generate a test signal that consists of two sine waves with frequencies of 10 Hz and 20 Hz. We then plot the original signal to visualize it.

Next, we perform the Wavelet transform on the signal using the pywt.wavedec function. This decomposes the signal into different frequency sub-bands, with the lowest frequencies in the first element of the output tuple, and the highest frequencies in the last element.

We then set a compression threshold and set all coefficients with absolute values below this threshold to zero using the pywt.threshold function. In this example, we set the threshold to 10% of the maximum coefficient in each sub-band.

Finally, we reconstruct the signal using the compressed coefficients using the pywt.waverec function. We plot the reconstructed signal to visualize the effect of the compression.

This is just a simple example, and in practice, there are many ways to choose the compression threshold and how to encode the compressed coefficients for efficient storage or transmission.